Inline imports in `conda.cli.main_*` and break up `conda.cli.conda_argparse` (#13173)
* Test all import paths for conda.cli.conda_argparse

* Remove F821 lint ignore

* Inline imports for main_clean

* Inline imports for main_compare

* Inline imports for main_compare

* Inline imports for main_create

* Inline import for main_info

* Inline import for main_init

* Inline import for main_install

* Inline import for main_list

* Inline import for main_notices

* Inline import for main_package

* Inline import for main_pip

* Inline import for main_remove

* Inline import for main_rename

* Inline import for main_run

* Inline imports for main_search

* Inline imports for main_update

* Move each subcommand's parser into main_*

* Remove redundant calls to add_parser_help

* Correct return code

* Fix tests

* Ruff

* Fix failures

* Cleanup

* Update new (de)activate mock modules

* Change comment to docstring


---------
