+++ conda/cli/main_info.py
@@ -4,29 +4,113 @@
 
 Display information about current conda installation.
 """
+from __future__ import annotations
+
 import json
 import os
 import re
 import sys
+from argparse import (
+    SUPPRESS,
+    ArgumentParser,
+    Namespace,
+    _StoreTrueAction,
+    _SubParsersAction,
+)
 from logging import getLogger
 from os.path import exists, expanduser, isfile, join
+from textwrap import wrap
+from typing import Iterable
 
-from .. import CONDA_PACKAGE_ROOT
-from .. import __version__ as conda_version
-from ..base.context import context, env_name, sys_rc_path, user_rc_path
-from ..common.compat import on_win
-from ..common.url import mask_anaconda_token
-from ..core.index import _supplement_index_with_system
 from ..deprecations import deprecated
-from ..models.channel import all_channel_urls, offline_keep
-from ..models.match_spec import MatchSpec
-from ..utils import human_bytes
-from .common import print_envs_list, stdout_json
 
 log = getLogger(__name__)
 
 
+def configure_parser(sub_parsers: _SubParsersAction, **kwargs) -> ArgumentParser:
+    from ..common.constants import NULL
+    from .helpers import add_parser_json
+
+    summary = "Display information about current conda install."
+    description = summary
+    epilog = ""
+
+    p = sub_parsers.add_parser(
+        "info",
+        help=summary,
+        description=description,
+        epilog=epilog,
+        **kwargs,
+    )
+    add_parser_json(p)
+    p.add_argument(
+        "--offline",
+        action="store_true",
+        default=NULL,
+        help=SUPPRESS,
+    )
+    p.add_argument(
+        "-a",
+        "--all",
+        dest="verbosity",
+        action=deprecated.action(
+            "24.3",
+            "24.9",
+            _StoreTrueAction,
+            addendum="Use `--verbose` instead.",
+        ),
+    )
+    p.add_argument(
+        "--base",
+        action="store_true",
+        help="Display base environment path.",
+    )
+    # TODO: deprecate 'conda info --envs' and create 'conda list --envs'
+    p.add_argument(
+        "-e",
+        "--envs",
+        action="store_true",
+        help="List all known conda environments.",
+    )
+    p.add_argument(
+        "-l",
+        "--license",
+        action="store_true",
+        help=SUPPRESS,
+    )
+    p.add_argument(
+        "-s",
+        "--system",
+        action="store_true",
+        help="List environment variables.",
+    )
+    p.add_argument(
+        "--root",
+        action="store_true",
+        help=SUPPRESS,
+        dest="base",
+    )
+    p.add_argument(
+        "--unsafe-channels",
+        action="store_true",
+        help="Display list of channels with tokens exposed.",
+    )
+
+    p.add_argument(
+        "packages",
+        action="store",
+        nargs="*",
+        help=SUPPRESS,
+    )
+
+    p.set_defaults(func="conda.cli.main_info.execute")
+
+    return p
+
+
 def get_user_site():  # pragma: no cover
+    from ..common.compat import on_win
+
     site_dirs = []
     try:
         if not on_win:
@@ -69,6 +153,8 @@ def dump_record(pkg):
 
 
 def pretty_package(prec):
+    from ..utils import human_bytes
+
     pkg = dump_record(prec)
     d = {
         "file name": prec.fn,
@@ -94,7 +180,11 @@ def pretty_package(prec):
 
 
 def print_package_info(packages):
+    from ..base.context import context
     from ..core.subdir_data import SubdirData
+    from ..deprecations import deprecated
+    from ..models.match_spec import MatchSpec
+    from .common import stdout_json
 
     results = {}
     for package in packages:
@@ -117,6 +207,14 @@ def print_package_info(packages):
 
 
 def get_info_dict(system=False):
+    from .. import CONDA_PACKAGE_ROOT
+    from .. import __version__ as conda_version
+    from ..base.context import context, env_name, sys_rc_path, user_rc_path
+    from ..common.compat import on_win
+    from ..common.url import mask_anaconda_token
+    from ..core.index import _supplement_index_with_system
+    from ..models.channel import all_channel_urls, offline_keep
+
     try:
         from requests import __version__ as requests_version
 
@@ -234,8 +332,6 @@ def get_info_dict(system=False):
 
 
 def get_env_vars_str(info_dict):
-    from textwrap import wrap
-
     builder = []
     builder.append("%23s:" % "environment variables")
     env_vars = info_dict.get("env_vars", {})
@@ -250,83 +346,68 @@ def get_env_vars_str(info_dict):
 
 
 def get_main_info_str(info_dict):
-    for key in "pkgs_dirs", "envs_dirs", "channels", "config_files":
-        info_dict[f"_{key}"] = ("\n" + 26 * " ").join(map(str, info_dict[key]))
+    from ..common.compat import on_win
 
-    info_dict["_virtual_pkgs"] = ("\n" + 26 * " ").join(
-        ["%s=%s=%s" % tuple(x) for x in info_dict["virtual_pkgs"]]
-    )
-    info_dict["_rtwro"] = "writable" if info_dict["root_writable"] else "read only"
+    def flatten(lines: Iterable[str]) -> str:
+        return ("\n" + 26 * " ").join(map(str, lines))
 
-    format_param = lambda nm, val: "%23s : %s" % (nm, val)
+    def builder():
+        if info_dict["active_prefix_name"]:
+            yield ("active environment", info_dict["active_prefix_name"])
+            yield ("active env location", info_dict["active_prefix"])
+        else:
+            yield ("active environment", info_dict["active_prefix"])
 
-    builder = [""]
+        if info_dict["conda_shlvl"] >= 0:
+            yield ("shell level", info_dict["conda_shlvl"])
 
-    if info_dict["active_prefix_name"]:
-        builder.append(
-            format_param("active environment", info_dict["active_prefix_name"])
+        yield ("user config file", info_dict["user_rc_path"])
+        yield ("populated config files", flatten(info_dict["config_files"]))
+        yield ("conda version", info_dict["conda_version"])
+        yield ("conda-build version", info_dict["conda_build_version"])
+        yield ("python version", info_dict["python_version"])
+        yield (
+            "virtual packages",
+            flatten("=".join(pkg) for pkg in info_dict["virtual_pkgs"]),
         )
-        builder.append(format_param("active env location", info_dict["active_prefix"]))
-    else:
-        builder.append(format_param("active environment", info_dict["active_prefix"]))
+        writable = "writable" if info_dict["root_writable"] else "read only"
+        yield ("base environment", f"{info_dict['root_prefix']}  ({writable})")
+        yield ("conda av data dir", info_dict["av_data_dir"])
+        yield ("conda av metadata url", info_dict["av_metadata_url_base"])
+        yield ("channel URLs", flatten(info_dict["channels"]))
+        yield ("package cache", flatten(info_dict["pkgs_dirs"]))
+        yield ("envs directories", flatten(info_dict["envs_dirs"]))
+        yield ("platform", info_dict["platform"])
+        yield ("user-agent", info_dict["user_agent"])
 
-    if info_dict["conda_shlvl"] >= 0:
-        builder.append(format_param("shell level", info_dict["conda_shlvl"]))
+        if on_win:
+            yield ("administrator", info_dict["is_windows_admin"])
+        else:
+            yield ("UID:GID", f"{info_dict['UID']}:{info_dict['GID']}")
 
-    builder.extend(
-        (
-            format_param("user config file", info_dict["user_rc_path"]),
-            format_param("populated config files", info_dict["_config_files"]),
-            format_param("conda version", info_dict["conda_version"]),
-            format_param("conda-build version", info_dict["conda_build_version"]),
-            format_param("python version", info_dict["python_version"]),
-            format_param("virtual packages", info_dict["_virtual_pkgs"]),
-            format_param(
-                "base environment",
-                "{}  ({})".format(info_dict["root_prefix"], info_dict["_rtwro"]),
-            ),
-            format_param("conda av data dir", info_dict["av_data_dir"]),
-            format_param("conda av metadata url", info_dict["av_metadata_url_base"]),
-            format_param("channel URLs", info_dict["_channels"]),
-            format_param("package cache", info_dict["_pkgs_dirs"]),
-            format_param("envs directories", info_dict["_envs_dirs"]),
-            format_param("platform", info_dict["platform"]),
-            format_param("user-agent", info_dict["user_agent"]),
-        )
-    )
+        yield ("netrc file", info_dict["netrc_file"])
+        yield ("offline mode", info_dict["offline"])
 
-    if on_win:
-        builder.append(format_param("administrator", info_dict["is_windows_admin"]))
-    else:
-        builder.append(
-            format_param("UID:GID", "{}:{}".format(info_dict["UID"], info_dict["GID"]))
-        )
-
-    builder.extend(
-        (
-            format_param("netrc file", info_dict["netrc_file"]),
-            format_param("offline mode", info_dict["offline"]),
-        )
-    )
-
-    builder.append("")
-    return "\n".join(builder)
+    return "\n".join(("", *(f"{key:>23} : {value}" for key, value in builder()), ""))
 
 
-def execute(args, parser):
+def execute(args: Namespace, parser: ArgumentParser) -> int:
+    from ..base.context import context
+    from .common import print_envs_list, stdout_json
+
     if args.base:
         if context.json:
             stdout_json({"root_prefix": context.root_prefix})
         else:
             print(f"{context.root_prefix}")
-        return
+        return 0
 
     if args.packages:
         from ..resolve import ResolvePackageNotFound
 
         try:
             print_package_info(args.packages)
-            return
+            return 0
         except ResolvePackageNotFound as e:  # pragma: no cover
             from ..exceptions import PackagesNotFoundError
 
@@ -383,3 +464,4 @@ def execute(args, parser):
 
     if context.json:
         stdout_json(info_dict)
+    return 0
