+++ conda/cli/main_remove.py
@@ -5,27 +5,144 @@
 Removes the specified packages from an existing environment.
 """
 import logging
+from argparse import ArgumentParser, Namespace, _SubParsersAction
 from os.path import isfile, join
-
-from ..base.context import context
-from ..core.envs_manager import unregister_env
-from ..core.link import PrefixSetup, UnlinkLinkTransaction
-from ..core.prefix_data import PrefixData
-from ..exceptions import (
-    CondaEnvironmentError,
-    CondaValueError,
-    DirectoryNotACondaEnvironmentError,
-    PackagesNotFoundError,
-)
-from ..gateways.disk.delete import path_is_clean, rm_rf
-from ..models.match_spec import MatchSpec
-from .common import check_non_admin, specs_from_args
-from .install import handle_txn
 
 log = logging.getLogger(__name__)
 
 
-def execute(args, parser):
+def configure_parser(sub_parsers: _SubParsersAction, **kwargs) -> ArgumentParser:
+    from ..auxlib.ish import dals
+    from ..common.constants import NULL
+    from .actions import NullCountAction
+    from .helpers import (
+        add_output_and_prompt_options,
+        add_parser_channels,
+        add_parser_networking,
+        add_parser_prefix,
+        add_parser_prune,
+        add_parser_pscheck,
+        add_parser_solver,
+    )
+
+    summary = "Remove a list of packages from a specified conda environment. "
+    description = dals(
+        f"""
+        {summary}
+
+        Use `--all` flag to remove all packages and the environment itself.
+
+        This command will also remove any package that depends on any of the
+        specified packages as well---unless a replacement can be found without
+        that dependency. If you wish to skip this dependency checking and remove
+        just the requested packages, add the '--force' option. Note however that
+        this may result in a broken environment, so use this with caution.
+        """
+    )
+    epilog = dals(
+        """
+        Examples:
+
+        Remove the package 'scipy' from the currently-active environment::
+
+            conda remove scipy
+
+        Remove a list of packages from an environemnt 'myenv'::
+
+            conda remove -n myenv scipy curl wheel
+
+        Remove all packages from environment `myenv` and the environment itself::
+
+            conda remove -n myenv --all
+
+        """
+    )
+
+    p = sub_parsers.add_parser(
+        "remove",
+        help=summary,
+        description=description,
+        epilog=epilog,
+        **kwargs,
+    )
+    add_parser_pscheck(p)
+
+    add_parser_prefix(p)
+    add_parser_channels(p)
+
+    solver_mode_options = p.add_argument_group("Solver Mode Modifiers")
+    solver_mode_options.add_argument(
+        "--all",
+        action="store_true",
+        help="Remove all packages, i.e., the entire environment.",
+    )
+    solver_mode_options.add_argument(
+        "--features",
+        action="store_true",
+        help="Remove features (instead of packages).",
+    )
+    solver_mode_options.add_argument(
+        "--force-remove",
+        "--force",
+        action="store_true",
+        help="Forces removal of a package without removing packages that depend on it. "
+        "Using this option will usually leave your environment in a broken and "
+        "inconsistent state.",
+        dest="force_remove",
+    )
+    solver_mode_options.add_argument(
+        "--no-pin",
+        action="store_true",
+        dest="ignore_pinned",
+        default=NULL,
+        help="Ignore pinned package(s) that apply to the current operation. "
+        "These pinned packages might come from a .condarc file or a file in "
+        "<TARGET_ENVIRONMENT>/conda-meta/pinned.",
+    )
+    add_parser_prune(solver_mode_options)
+    add_parser_solver(solver_mode_options)
+
+    add_parser_networking(p)
+    add_output_and_prompt_options(p)
+
+    p.add_argument(
+        "package_names",
+        metavar="package_name",
+        action="store",
+        nargs="*",
+        help="Package names to remove from the environment.",
+    )
+    p.add_argument(
+        "--dev",
+        action=NullCountAction,
+        help="Use `sys.executable -m conda` in wrapper scripts instead of CONDA_EXE. "
+        "This is mainly for use during tests where we test new conda sources "
+        "against old Python versions.",
+        dest="dev",
+        default=NULL,
+    )
+
+    p.set_defaults(func="conda.cli.main_remove.execute")
+
+    return p
+
+
+def execute(args: Namespace, parser: ArgumentParser) -> int:
+    from ..base.context import context
+    from ..core.envs_manager import unregister_env
+    from ..core.link import PrefixSetup, UnlinkLinkTransaction
+    from ..core.prefix_data import PrefixData
+    from ..exceptions import (
+        CondaEnvironmentError,
+        CondaValueError,
+        DirectoryNotACondaEnvironmentError,
+        PackagesNotFoundError,
+    )
+    from ..gateways.disk.delete import path_is_clean, rm_rf
+    from ..models.match_spec import MatchSpec
+    from .common import check_non_admin, specs_from_args
+    from .install import handle_txn
+
     if not (args.all or args.package_names):
         raise CondaValueError(
             "no package names supplied,\n"
@@ -86,7 +203,7 @@ def execute(args, parser):
             rm_rf(prefix, clean_empty_parents=True)
             unregister_env(prefix)
 
-        return
+        return 0
 
     else:
         if args.features:
@@ -99,3 +216,4 @@ def execute(args, parser):
         solver = solver_backend(prefix, channel_urls, subdirs, specs_to_remove=specs)
         txn = solver.solve_for_transaction()
         handle_txn(txn, prefix, args, False, True)
+        return 0
