+++ tests/cli/test_conda_argparse.py
@@ -1,7 +1,10 @@
 # Copyright (C) 2012 Anaconda, Inc
 # SPDX-License-Identifier: BSD-3-Clause
+import importlib
 import re
+from inspect import isclass, isfunction
 from logging import getLogger
+from typing import Any, Callable
 
 import pytest
 
@@ -46,3 +49,61 @@ def test_cli_args_as_strings(conda_cli: CondaCLIFixture):
     assert stdout
     assert not stderr
     assert not err
+
+
+@pytest.mark.parametrize(
+    "path,validate",
+    [
+        ("conda.cli.conda_argparse.add_output_and_prompt_options", isfunction),
+        ("conda.cli.conda_argparse.add_parser_channels", isfunction),
+        ("conda.cli.conda_argparse.add_parser_create_install_update", isfunction),
+        ("conda.cli.conda_argparse.add_parser_default_packages", isfunction),
+        ("conda.cli.conda_argparse.add_parser_help", isfunction),
+        ("conda.cli.conda_argparse.add_parser_json", isfunction),
+        ("conda.cli.conda_argparse.add_parser_known", isfunction),
+        ("conda.cli.conda_argparse.add_parser_networking", isfunction),
+        ("conda.cli.conda_argparse.add_parser_package_install_options", isfunction),
+        ("conda.cli.conda_argparse.add_parser_prefix", isfunction),
+        ("conda.cli.conda_argparse.add_parser_prune", isfunction),
+        ("conda.cli.conda_argparse.add_parser_pscheck", isfunction),
+        ("conda.cli.conda_argparse.add_parser_show_channel_urls", isfunction),
+        ("conda.cli.conda_argparse.add_parser_solver", isfunction),
+        ("conda.cli.conda_argparse.add_parser_solver_mode", isfunction),
+        ("conda.cli.conda_argparse.add_parser_update_modifiers", isfunction),
+        ("conda.cli.conda_argparse.add_parser_verbose", isfunction),
+        # derived from argparse.ArgumentParser
+        ("conda.cli.conda_argparse.ArgumentParser", isclass),
+        ("conda.cli.conda_argparse.BUILTIN_COMMANDS", lambda x: isinstance(x, set)),
+        ("conda.cli.conda_argparse.configure_parser_clean", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_compare", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_config", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_create", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_info", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_init", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_install", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_list", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_notices", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_package", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_plugins", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_remove", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_rename", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_run", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_search", isfunction),
+        ("conda.cli.conda_argparse.configure_parser_update", isfunction),
+        ("conda.cli.conda_argparse.do_call", isfunction),
+        ("conda.cli.conda_argparse.escaped_sys_rc_path", lambda x: isinstance(x, str)),
+        ("conda.cli.conda_argparse.escaped_user_rc_path", lambda x: isinstance(x, str)),
+        ("conda.cli.conda_argparse.ExtendConstAction", isclass),
+        ("conda.cli.conda_argparse.find_builtin_commands", isfunction),
+        ("conda.cli.conda_argparse.generate_parser", isfunction),
+        ("conda.cli.conda_argparse.generate_pre_parser", isfunction),
+        ("conda.cli.conda_argparse.NullCountAction", isclass),
+        ("conda.cli.conda_argparse.sys_rc_path", lambda x: isinstance(x, str)),
+        ("conda.cli.conda_argparse.user_rc_path", lambda x: isinstance(x, str)),
+    ],
+)
+def test_imports(path: str, validate: Callable[[Any], bool]):
+    path, attr = path.rsplit(".", 1)
+    module = importlib.import_module(path)
+    assert hasattr(module, attr)
+    assert validate(getattr(module, attr))
